+++ tests/style/bug_137/test.m
@@ -1,0 +1,8 @@
+
+
+
+% (c) Copyright 2020 Florian Schanda
+%
+% empty lines at the start of the file should be trimmed
+
+disp "potato";